+++ bin/gamedig.js
@@ -4,7 +4,7 @@ const Minimist = require('minimist'),
     Gamedig = require('..');
 
 const argv = Minimist(process.argv.slice(2), {
-    boolean: ['pretty','debug','givenPortOnly'],
+    boolean: ['pretty','debug','givenPortOnly','requestRules'],
     string: ['guildId','listenUdpPort']
 });